Improve library TLD data structure!  v0.1.1
Ensure library TLD data structure is based on a constant map, from
which we can define convenient constants (though they are not required
as one can use the map directly).

The interface, in future, may still be more similar to biter777's
countries library, so I will have a think about a nice API.

I need to write tests and usage examples so that I can think of a good
API to use.